# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Media Library API",
        "description": "API for managing book and movie rentals",
        "version": "1.0.0"
    },
    "paths": {
...
        "/getRentals": {
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get rentals by date",
                "operationId": "8a22d1a11039781c2ba5ef8365578e66",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Number of days to filter rentals",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of rentals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rental"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/createUser": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "91bb49006fe122575810ade6bfe22a18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "First name of the user",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Last name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email of the user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password of the user",
                                        "type": "string"
                                    },
                                    "user_type": {
                                        "description": "Type of user (student or teacher)",
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "Gender of the user",
                                        "type": "string"
                                    },
                                    "grade_level": {
                                        "description": "Grade level of the user (if student)",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "department": {
                                        "description": "Department of the user (if teacher)",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "operationId": "2c99028953aa155c0c0bd6aa03bbcd2d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email of the user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password of the user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "your_jwt_token_here"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid login credentials"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "title": "Book",
                "description": "Book model",
                "properties": {
                    "item_id": {
                        "description": "Unique identifier for the book",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the book",
                        "type": "string"
                    },
                    "author": {
                        "description": "Author of the book",
                        "type": "string"
                    },
                    "publication_year": {
                        "description": "Year the book was published",
                        "type": "integer"
                    },
                    "genre": {
                        "description": "Genre of the book",
                        "type": "string"
                    },
                    "availability": {
                        "description": "Availability status of the book (true if available, false if rented)",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Movie": {
                "title": "Movie",
                "description": "Movie model",
                "properties": {
                    "item_id": {
                        "description": "Unique identifier for the movie",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the movie",
                        "type": "string"
                    },
                    "director": {
                        "description": "Director of the movie",
                        "type": "string"
                    },
                    "publication_year": {
                        "description": "Year the movie was published",
                        "type": "integer"
                    },
                    "genre": {
                        "description": "Genre of the movie",
                        "type": "string"
                    },
                    "availability": {
                        "description": "Availability status of the movie (true if available, false if rented)",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Rental": {
                "title": "Rental",
                "description": "Rental model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the rental",
                        "type": "integer"
                    },
                    "renter_id": {
                        "description": "The ID of the user who rented the item",
                        "type": "integer"
                    },
                    "rental_date": {
                        "description": "The date the item was rented",
                        "type": "string",
                        "format": "date"
                    },
                    "return_date": {
                        "description": "The date the item was returned",
                        "type": "string",
                        "format": "date"
                    },
                    "returned": {
                        "description": "Status of whether the item has been returned",
                        "type": "boolean"
                    },
                    "book_id": {
                        "description": "ID of the rented book, null if a movie is rented",
                        "type": "integer",
                        "nullable": true
                    },
                    "movie_id": {
                        "description": "ID of the rented movie, null if a book is rented",
                        "type": "integer",
                        "nullable": true
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "movies": {
                        "$ref": "#/components/schemas/Movie"
                    },
                    "books": {
                        "$ref": "#/components/schemas/Book"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "user_id": {
                        "description": "Unique identifier for the user",
                        "type": "integer"
                    },
                    "first_name": {
                        "description": "First name of the user",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Last name of the user",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string",
                        "format": "email"
                    },
                    "user_type": {
                        "description": "Type of user, either 'student' or 'teacher'",
                        "type": "string"
                    },
                    "grade_level": {
                        "description": "Grade level of the user (applicable if the user is a student)",
                        "type": "integer",
                        "nullable": true
                    },
                    "department": {
                        "description": "Department of the user (applicable if the user is a teacher)",
                        "type": "string",
                        "nullable": true
                    },
                    "gender": {
                        "description": "Gender of the user",
                        "type": "string"
                    },
                    "rentals": {
                        "description": "List of rentals associated with the user",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rental"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}